pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["Drsultanali <sultan_ali338@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.scripts]
dev = "todos.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
